fix dump config

+++ esphome/components/adc/adc_sensor_esp32.cpp
@@ -78,7 +78,7 @@ void ADCSensor::dump_config() {
     }
   }
   ESP_LOGCONFIG(TAG, "  Samples: %i", this->sample_count_);
-  ESP_LOGCONFIG(TAG, "  Sampling mode: %i", this->sampling_mode_);
+  ESP_LOGCONFIG(TAG, "  Sampling mode: %hhu", static_cast<uint8_t>(this->sampling_mode_));
   LOG_UPDATE_INTERVAL(this);
 }
