+++ esphome/components/adc/adc_sensor_libretiny.cpp
@@ -23,7 +23,7 @@ void ADCSensor::dump_config() {
   LOG_PIN("  Pin: ", this->pin_);
 #endif  // USE_ADC_SENSOR_VCC
   ESP_LOGCONFIG(TAG, "  Samples: %i", this->sample_count_);
-  ESP_LOGCONFIG(TAG, "  Sampling mode: %i", this->sampling_mode_);
+  ESP_LOGCONFIG(TAG, "  Sampling mode: %hhu", static_cast<uint8_t>(this->sampling_mode_));
   LOG_UPDATE_INTERVAL(this);
 }
