+++ esphome/components/adc/adc_sensor_rp2040.cpp
@@ -34,7 +34,7 @@ void ADCSensor::dump_config() {
 #endif  // USE_ADC_SENSOR_VCC
   }
   ESP_LOGCONFIG(TAG, "  Samples: %i", this->sample_count_);
-  ESP_LOGCONFIG(TAG, "  Sampling mode: %i", this->sampling_mode_);
+  ESP_LOGCONFIG(TAG, "  Sampling mode: %hhu", static_cast<uint8_t>(this->sampling_mode_));
   LOG_UPDATE_INTERVAL(this);
 }
